refactor: HandlerMapping, HandlerAdapter 등록 로직을 AppWebApplicationInitializer에서 관리하도록 변경
