+++ app/src/main/java/com/techcourse/Application.java
@@ -1,12 +1,11 @@
 package com.techcourse;
 
+import java.io.File;
+import java.util.stream.Stream;
 import org.apache.catalina.connector.Connector;
 import org.apache.catalina.startup.Tomcat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.File;
-import java.util.stream.Stream;
 
 public class Application {
