+++ app/src/test/java/com/techcourse/ManualHandlerAdapterTest.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package com.techcourse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.when;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
